+++ packages/account/tests/account.spec.ts
@@ -1,19 +1,17 @@
-
-import hardhat from 'hardhat'
-import * as chai from 'chai'
-import * as utils from '@0xsequence/tests'
-
-import { ethers } from 'ethers'
-import { Orchestrator } from '@0xsequence/signhub'
-import { Account } from '../src/account'
+import { commons, v1, v2 } from '@0xsequence/core'
 import { migrator } from '@0xsequence/migration'
 import { NetworkConfig } from '@0xsequence/network'
-import { tracker, trackers } from '@0xsequence/sessions'
 import { LocalRelayer, Relayer } from '@0xsequence/relayer'
-import { commons, v1, v2 } from '@0xsequence/core'
-import chaiAsPromised from 'chai-as-promised'
+import { tracker, trackers } from '@0xsequence/sessions'
+import { Orchestrator } from '@0xsequence/signhub'
+import * as utils from '@0xsequence/tests'
 import { Wallet } from '@0xsequence/wallet'
-import { encodeBundleExecData } from '@0xsequence/core/src/commons/transaction'
+import * as chai from 'chai'
+import chaiAsPromised from 'chai-as-promised'
+import { ethers } from 'ethers'
+import hardhat from 'hardhat'
+
+import { Account } from '../src/account'
 
 const { expect } = chai.use(chaiAsPromised)
 
@@ -1034,7 +1032,7 @@ describe('Account', () => {
           const deployTx = Wallet.buildDeployTransaction(contexts[1], imageHash)
           await signer1.sendTransaction({
             to: deployTx.entrypoint,
-            data: encodeBundleExecData(deployTx),
+            data: commons.transaction.encodeBundleExecData(deployTx),
           })
   
           expect(await networks[0].provider!.getCode(account.address).then((c) => ethers.utils.arrayify(c).length))
@@ -1056,7 +1054,7 @@ describe('Account', () => {
           const deployTx = Wallet.buildDeployTransaction(contexts[1], imageHash)
           await signer1.sendTransaction({
             to: deployTx.entrypoint,
-            data: encodeBundleExecData(deployTx),
+            data: commons.transaction.encodeBundleExecData(deployTx),
           })
   
           expect(await networks[0].provider!.getCode(account.address).then((c) => ethers.utils.arrayify(c).length))
@@ -1071,7 +1069,7 @@ describe('Account', () => {
           const deployTx = Wallet.buildDeployTransaction(contexts[1], imageHash)
           await signer1.sendTransaction({
             to: deployTx.entrypoint,
-            data: encodeBundleExecData(deployTx),
+            data: commons.transaction.encodeBundleExecData(deployTx),
           })
   
           expect(await networks[0].provider!.getCode(account.address).then((c) => ethers.utils.arrayify(c).length))