+++ packages/network/src/constants.ts
@@ -1,0 +1,343 @@
+export enum ChainId {
+  // Ethereum
+  MAINNET = 1,
+  ROPSTEN = 3,
+  RINKEBY = 4,
+  GOERLI = 5,
+  KOVAN = 42,
+
+  // Polygon
+  POLYGON = 137,
+  POLYGON_MUMBAI = 80001,
+  POLYGON_ZKEVM = 1101,
+
+  // BSC
+  BSC = 56,
+  BSC_TESTNET = 97,
+
+  // Optimism
+  OPTIMISM = 10,
+  OPTIMISM_TESTNET = 69,
+
+  // Arbitrum One
+  ARBITRUM = 42161,
+  ARBITRUM_GOERLI = 421613,
+
+  // Arbitrum Nova
+  ARBITRUM_NOVA = 42170,
+
+  // Avalanche
+  AVALANCHE = 43114,
+  AVALANCHE_TESTNET = 43113,
+
+  // Fantom
+  FANTOM = 250,
+  FANTOM_TESTNET = 4002,
+
+  // Gnosis Chain (XDAI)
+  GNOSIS = 100,
+
+  // AURORA
+  AURORA = 1313161554,
+  AURORA_TESTNET = 1313161556,
+
+  // BASE
+  BASE_GOERLI = 84531,
+
+  // HOMEVERSE
+  HOMEVERSE_TESTNET = 40875,
+
+  // HARDHAT TESTNETS
+  HARDHAT = 31337,
+  HARDHAT_2 = 31338
+}
+
+export enum NetworkType {
+  MAINNET = 'mainnet',
+  TESTNET = 'testnet'
+}
+
+export type BlockExplorerConfig = {
+  name?: string
+  rootUrl: string
+  addressUrl?: string
+  txnHashUrl?: string
+}
+
+export interface NetworkMetadata {
+  chainId: ChainId
+  type?: NetworkType
+  name: string
+  title?: string
+  blockExplorer?: BlockExplorerConfig
+  ensAddress?: string
+  testnet?: boolean // Deprecated
+}
+
+export const networks: Record<ChainId, NetworkMetadata> = {
+  [ChainId.MAINNET]: {
+    chainId: ChainId.MAINNET,
+    type: NetworkType.MAINNET,
+    name: 'mainnet',
+    title: 'Ethereum',
+    blockExplorer: {
+      name: 'Etherscan',
+      rootUrl: 'https://etherscan.io/'
+    },
+    ensAddress: '0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e'
+  },
+  [ChainId.ROPSTEN]: {
+    chainId: ChainId.ROPSTEN,
+    type: NetworkType.TESTNET,
+    name: 'ropsten',
+    title: 'Ropsten',
+    testnet: true,
+    blockExplorer: {
+      name: 'Etherscan (Ropsten)',
+      rootUrl: 'https://ropsten.etherscan.io/'
+    },
+    ensAddress: '0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e'
+  },
+  [ChainId.RINKEBY]: {
+    chainId: ChainId.RINKEBY,
+    type: NetworkType.TESTNET,
+    name: 'rinkeby',
+    title: 'Rinkeby',
+    testnet: true,
+    blockExplorer: {
+      name: 'Etherscan (Rinkeby)',
+      rootUrl: 'https://rinkeby.etherscan.io/'
+    },
+    ensAddress: '0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e'
+  },
+  [ChainId.GOERLI]: {
+    chainId: ChainId.GOERLI,
+    type: NetworkType.TESTNET,
+    name: 'goerli',
+    title: 'Goerli',
+    testnet: true,
+    blockExplorer: {
+      name: 'Etherscan (Goerli)',
+      rootUrl: 'https://goerli.etherscan.io/'
+    },
+    ensAddress: '0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e'
+  },
+  [ChainId.KOVAN]: {
+    chainId: ChainId.KOVAN,
+    type: NetworkType.TESTNET,
+    name: 'kovan',
+    title: 'Kovan',
+    testnet: true,
+    blockExplorer: {
+      name: 'Etherscan (Kovan)',
+      rootUrl: 'https://kovan.etherscan.io/'
+    }
+  },
+  [ChainId.POLYGON]: {
+    chainId: ChainId.POLYGON,
+    type: NetworkType.MAINNET,
+    name: 'polygon',
+    title: 'Polygon',
+    blockExplorer: {
+      name: 'Polygonscan',
+      rootUrl: 'https://polygonscan.com/'
+    }
+  },
+  [ChainId.POLYGON_MUMBAI]: {
+    chainId: ChainId.POLYGON_MUMBAI,
+    type: NetworkType.TESTNET,
+    name: 'mumbai',
+    title: 'Polygon Mumbai',
+    testnet: true,
+    blockExplorer: {
+      name: 'Polygonscan (Mumbai)',
+      rootUrl: 'https://mumbai.polygonscan.com/'
+    }
+  },
+  [ChainId.POLYGON_ZKEVM]: {
+    chainId: ChainId.POLYGON_ZKEVM,
+    type: NetworkType.MAINNET,
+    name: 'polygon-zkevm',
+    title: 'Polygon zkEVM',
+    blockExplorer: {
+      name: 'Polygonscan (zkEVM)',
+      rootUrl: 'https://zkevm.polygonscan.com/'
+    }
+  },
+  [ChainId.BSC]: {
+    chainId: ChainId.BSC,
+    type: NetworkType.MAINNET,
+    name: 'bsc',
+    title: 'BNB Smart Chain',
+    blockExplorer: {
+      name: 'BSCScan',
+      rootUrl: 'https://bscscan.com/'
+    }
+  },
+  [ChainId.BSC_TESTNET]: {
+    chainId: ChainId.BSC_TESTNET,
+    type: NetworkType.TESTNET,
+    name: 'bsc-testnet',
+    title: 'BNB Smart Chain Testnet',
+    testnet: true,
+    blockExplorer: {
+      name: 'BSCScan (Testnet)',
+      rootUrl: 'https://testnet.bscscan.com/'
+    }
+  },
+  [ChainId.OPTIMISM]: {
+    chainId: ChainId.OPTIMISM,
+    type: NetworkType.MAINNET,
+    name: 'optimism',
+    title: 'Optimism',
+    blockExplorer: {
+      name: 'Etherscan (Optimism)',
+      rootUrl: 'https://optimistic.etherscan.io/'
+    }
+  },
+  [ChainId.OPTIMISM_TESTNET]: {
+    chainId: ChainId.OPTIMISM_TESTNET,
+    type: NetworkType.TESTNET,
+    name: 'optimism-testnet',
+    title: 'Optimistic Kovan',
+    testnet: true,
+    blockExplorer: {
+      name: 'Etherscan (Optimism Testnet)',
+      rootUrl: 'https://kovan-optimistic.etherscan.io/'
+    }
+  },
+  [ChainId.ARBITRUM]: {
+    chainId: ChainId.ARBITRUM,
+    type: NetworkType.MAINNET,
+    name: 'arbitrum',
+    title: 'Arbitrum One',
+    blockExplorer: {
+      name: 'Arbiscan',
+      rootUrl: 'https://arbiscan.io/'
+    }
+  },
+  [ChainId.ARBITRUM_GOERLI]: {
+    chainId: ChainId.ARBITRUM_GOERLI,
+    type: NetworkType.TESTNET,
+    name: 'arbitrum-goerli',
+    title: 'Arbitrum Goerli',
+    testnet: true,
+    blockExplorer: {
+      name: 'Arbiscan (Testnet)',
+      rootUrl: 'https://testnet.arbiscan.io/'
+    }
+  },
+  [ChainId.ARBITRUM_NOVA]: {
+    chainId: ChainId.ARBITRUM_NOVA,
+    type: NetworkType.MAINNET,
+    name: 'arbitrum-nova',
+    title: 'Arbitrum Nova',
+    blockExplorer: {
+      name: 'Arbiscan Nova',
+      rootUrl: 'https://nova.arbiscan.io/'
+    }
+  },
+  [ChainId.AVALANCHE]: {
+    chainId: ChainId.AVALANCHE,
+    type: NetworkType.MAINNET,
+    name: 'avalanche',
+    title: 'Avalanche',
+    blockExplorer: {
+      name: 'Snowtrace',
+      rootUrl: 'https://snowtrace.io/'
+    }
+  },
+  [ChainId.AVALANCHE_TESTNET]: {
+    chainId: ChainId.AVALANCHE_TESTNET,
+    type: NetworkType.TESTNET,
+    name: 'avalanche-testnet',
+    title: 'Avalanche Testnet',
+    testnet: true,
+    blockExplorer: {
+      name: 'Snowtrace (Testnet)',
+      rootUrl: 'https://testnet.snowtrace.io/'
+    }
+  },
+  [ChainId.FANTOM]: {
+    chainId: ChainId.FANTOM,
+    type: NetworkType.MAINNET,
+    name: 'fantom',
+    title: 'Fantom',
+    blockExplorer: {
+      name: 'FTMScan',
+      rootUrl: 'https://ftmscan.com/'
+    }
+  },
+  [ChainId.FANTOM_TESTNET]: {
+    chainId: ChainId.FANTOM_TESTNET,
+    type: NetworkType.TESTNET,
+    name: 'fantom-testnet',
+    title: 'Fantom Testnet',
+    testnet: true,
+    blockExplorer: {
+      name: 'FTMScan (Testnet)',
+      rootUrl: 'https://testnet.ftmscan.com/'
+    }
+  },
+  [ChainId.GNOSIS]: {
+    chainId: ChainId.GNOSIS,
+    type: NetworkType.MAINNET,
+    name: 'gnosis',
+    title: 'Gnosis Chain',
+    blockExplorer: {
+      name: 'Gnosis Chain Explorer',
+      rootUrl: 'https://blockscout.com/xdai/mainnet/'
+    }
+  },
+  [ChainId.AURORA]: {
+    chainId: ChainId.AURORA,
+    type: NetworkType.MAINNET,
+    name: 'aurora',
+    title: 'Aurora',
+    blockExplorer: {
+      name: 'Aurora Explorer',
+      rootUrl: 'https://aurorascan.dev/'
+    }
+  },
+  [ChainId.AURORA_TESTNET]: {
+    chainId: ChainId.AURORA_TESTNET,
+    type: NetworkType.TESTNET,
+    name: 'aurora-testnet',
+    title: 'Aurora Testnet',
+    blockExplorer: {
+      name: 'Aurora Explorer (Testnet)',
+      rootUrl: 'https://testnet.aurorascan.dev/'
+    }
+  },
+  [ChainId.BASE_GOERLI]: {
+    chainId: ChainId.BASE_GOERLI,
+    type: NetworkType.TESTNET,
+    name: 'base-goerli',
+    title: 'Base Goerli',
+    blockExplorer: {
+      name: 'Base Goerli Explorer',
+      rootUrl: 'https://goerli.basescan.org/'
+    }
+  },
+  [ChainId.HOMEVERSE_TESTNET]: {
+    chainId: ChainId.HOMEVERSE_TESTNET,
+    type: NetworkType.TESTNET,
+    name: 'homeverse-testnet',
+    title: 'Oasys Homeverse Testnet',
+    blockExplorer: {
+      name: 'Oasys Homeverse Explorer (Testnet)',
+      rootUrl: 'https://explorer.testnet.oasys.homeverse.games/'
+    }
+  },
+
+  [ChainId.HARDHAT]: {
+    chainId: ChainId.HARDHAT,
+    name: 'hardhat',
+    title: 'Hardhat (local testnet)'
+  },
+  [ChainId.HARDHAT_2]: {
+    chainId: ChainId.HARDHAT_2,
+    name: 'hardhat2',
+    title: 'Hardhat (local testnet)'
+  }
+}