+++ packages/guard/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @0xsequence/guard
 
+## 1.0.5
+
+### Patch Changes
+
+- network: export network constants
+- guard: use the correct global for fetch
+- network: nova-explorer.arbitrum.io -> nova.arbiscan.io
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @0xsequence/core@1.0.5
+  - @0xsequence/signhub@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes